Allow stinking clouds to spread over liquids
This was mainly relevant on the Plane of Fire, where the stinking clouds
were weirdly coming out of the lava (flavor-wise) but couldn't actually
be on the lava.

Also indirectly now allows players to actually center stinking clouds
on liquid, which makes plenty of sense.

Downside to this is that Plane of Fire stinking clouds are mostly back
to being their old diamond-shaped selves. Perhaps there should be a
small random component in how the cloud spreads.