Refactor: make create_rooms specific x and y use absolute coordinates
The level-relative code that applied numbers 1-5 in fifths of the level
was very weird and clunky, but more importantly it was going to get in
the way of planned level generator changes, which would require an
absolute coordinate to be specified.

create_rooms will now interpret x and y, if non-negative, as absolute
coordinates (otherwise it'll pick randomly). This means that any special
levels that specify x and y will get messed up. Luckily, the only levels
that did this were several Minetown variants. So, those are also changed
to being absolutely positioned.

A possible side effect: since vaults no longer have to be constrained to
a psuedo-fifth of the map, they may become more common.