Initial work on maze level rooms
Editing the maze creation routines to slap down some rooms before
walking the maze. Rooms currently fit nicely into the maze; they are
guaranteed to be aligned properly with it. They use the existing mkroom
structures, so rooms[] and roomno and everything else work properly.

The algorithm for placing rooms does not use rects and is rather simpler
in my opinion. I may later try applying this algorithm to the regular
dungeon generator.

Rooms currently usually do not have doors, but if the dead-end-removal
code from 3.6.1 happens to get called from a maze dead end into a room,
that code has been changed to pick up on it and create a door there.
More doors will be added soon.

This disables some of the 3.6.1 maze features as well; wide-corridor and
wide-wall mazes, and corrmazes are currently turned off entirely.
Possibly, corrmazes are OK to put back in, I will try later, but I think
widemazes will have to stay removed.