Fix: mazewalk would sometimes fail completely
This was on account of it occasionally picking an initial coordinate in
a room and not on its edge; having no STONE to reach from there, it
couldn't do anything.

Now, keep retrying coordinates until it finds something not in a room.
(The level should never be completely filled with rooms.)