Add "unfair-scare-less" conduct
This conduct tracks whether you have ever scared a monster by standing
on a space that made you immune to its attacks - anything that onscary()
returns true for is fair game; the main causes are Elbereth, the scroll
of scare monster, and sanctuary from being inside one of your own
temples.

You are permitted to write Elbereth all you want, so long as you never
actually scare anything with it.

Currently, other methods of scaring monsters, such as cause fear and
tooled horns, are permitted. This may be changed in the future and the
conduct will become never intentionally scaring a monster (it is hard to
get around scaring via reducing monsters to low hitpoints or shattering
their weapon).

The conduct is only broken when a non-fleeing monster gets scared and
starts fleeing. It is possible for a fleeing monster to silently get
scared again and keep fleeing, which is possibly abusable, but I think
it's best and most consistent to say that the conduct is broken when
"The monster turns to flee" because of where you're standing.