Fix: player was prompted to confirm trap move while still stuck
Mis-ordering of the logic meant that the confirmation prompt for moving
into a trap popped up before it realized you're still stuck in a trap
and can't move there in the first place.

There is a vanilla bug in which moving between two adjacent pits will
trigger the pit damage (without prompting for paranoid_trap). I'm not
addressing that since I reported it and it should eventually be fixed.