Allow forcefighting of webs with a bladed weapon to destroy them
May fail, based on a Dexterity check. Non-edged weapons don't work. If
you are wielding Sting, it always succeeds.

Also buff Sting when moving out of webs: it now destroys the web
outright. There is a minor exploit allowing you to destroy a web you're
stuck in in 0 turns; see the first FIXME.

Also, there are a couple edge cases that could get weird with
forcefighting a web that also has a remembered unseen monster on it -
see the second FIXME for details.