Fix: jumping over water didn't work
This was actually not due to paranoid water behavior as I had suspected;
it was because of the auto-ID of water walking boots change and its
drown() refactor. The check in hurtle_step() for whether to call drown()
and abort had been removed, which actually prevented *any* hurtling over
water.

This fixes it by re-adding those checks to hurtle_step(); this shouldn't
hurt that refactor.