Replace armor and weapons with materials-neutral versions
These are the objects that either:

1) Had a specific material in their unidentified name, such as iron
skull caps. This is the easiest one to change: just edit the
unidentified name to be material-agnostic.

2) Had a specific material in their regular name, like leather armors.
This affected the object constant so required those constants be changed
throughout the code. These are:
  leather armor -> light armor
  studded leather armor -> studded armor
  leather jacket -> jacket
  leather gloves -> gloves
  leather cloak -> plain cloak
  elven leather helm -> elven helm (with no more unIDed appearance)
  elven mithril-coat -> elven chain mail (now copper)
  dwarvish mithril-coat -> dwarvish chain mail (now iron)
  iron shoes -> dwarvish boots
  silver saber -> saber

  The one item which is NOT changed, and whose material won't be
  randomized anyway, is crystal plate mail. Still figuring out how to
  handle that.

3) Existed solely as an alternate-material counterpart to another object
that was otherwise identical in stats, such as silver daggers. These are
removed outright. Probabilities for counterpart weapons are increased by
the amount that their deleted counterpart had.

Also fix up tiles, des files, etc to reflect edited or removed objects.

This does not change objects with explicit materials but which won't
randomize those materials - such as rings, wands, etc.