Artifacts always have one specific material
For certain artifacts, this is a specific non-base material: Sunsword is
gold, Grayswandir, Werebane, and Demonbane are silver, and the Express
Card is platinum.

For all other artifacts, their material is the material of their base
object and cannot be anything different.

This is set at the time the object receives its name, which means that
if we wanted to do something like make Sting be metal, it would always
turn from wood into metal when the player named it.