Allow wishing for object materials in wizard mode
Initially I had this as wishing for any valid object material on a
non-artifact object outside of wizard mode, but I've hidden it behind an
 #if 0 for now. The reason for this is that I'd like to see how the game
plays out without being able to plan your ascension kit around wishing
up certain materials on objects.

Within wizard mode, the validity check is suppressed - you can wish for
any material and get it even if it doesn't make sense. Use this power
wisely.

Also contains a refactor of the wishing for gold code due to a necessary
change to accommodate that wishing for "500 gold" would parse the gold
as a material. Now, instead of creating the gold outright and returning,
readobjnam just sets typ to GOLD_PIECE and jumps forward to create it.