Add weight and AC differences for objects of different materials
Weights are computed via multiplying the base weight of an object by its
specific gravity (the fraction of its material's density over the
density of the regular material for that object). I may have gone a
little too realistic here; since specific gravity is dimensionless that
allows us to define material densities in any units we like, so I used
SI kg/m^3 and tried to approximate as close to real-life ones as
possible. This may have some undesirable gameplay effects such as a
platinum sword being over two and a half times as heavy as its regular
iron counterpart. We'll see.

AC differences are more arbitrary, and the relative differences in AC
between two materials are mostly just copied from GruntHack, which did
this first. The only change from that is that mithril grants an extra
point of AC compared to iron.