Silver gauntlets and boots deal silver damage to monsters
Also make monsters hitting with silver gauntlets or kicking with silver
boots deal silver damage to you.

I suppose touch attacks against silver body armor are another candidate
for implementing, but that's probably unbalanced given how many of
demons' attacks are weapon or claw attacks, and they are not present in
this commit.

Also fix an odd asymmetry in which getting hit by a silver weapon only
abused Con and did not apply the extra 20 points of silver damage.