Implement monster introducing from SpliceHack
Nymphs and foocubi (or more specifically anything with AD_SEDU and
AD_SSEX attacks) will now call mintroduce() which causes the monster to
introduce itself to you with a randomly chosen name. Name lists are
different for nymphs, male demons, and female demons, and are not the
same as in SpliceHack.

Nymphs won't introduce if they go through the "bragging about loot"
routine; they'll only do it when they at least attempt to charm you.