Port permahallu and permadeaf roleplay options from SpliceHack
These behave the same way as permanent blindness; you are hallucinating
or deaf from turn 1 and remain that way for the entire rest of the game.
They are set in the config with OPTIONS=hallucinating and OPTIONS=deaf.

Permanent hallucination can only be broken (permanently; it doesn't
come back) by equipping an artifact that confers hallucination
resistance (currently only Grayswandir). Permanent deafness cannot
currently be broken.

They are encoded in the xlogfile's achieve field.