New bogusmons, engravings, rumors, headstones, shirts
bogusmon.txt is organized a bit better. Many of the new ones are from
SpliceHack.

engrave.txt is also organized a bit better. This and the non-bogusmon
files don't have quite as many things added to them.

It is probably worth refactoring shirt messages into a new txt file in
the future since there are so many of them now, but I'm not doing it at
this point.