Import Tangles' whereis patch
On every level change, this writes a file to VARDIR/whereis/ named
"playername.whereis", which contains a selection of xlogfile fields
describing where the player is in the dungeon. The main use case for
this is server play, so people can view where a player is in the game
without actually connecting to the server and spectating.

This creates the whereis folder in the top level Makefile under the
install target, so it is not necessary to special case creating it in
other cases.