Golems may drop other items of the appropriate material
The addition of object materials allows this to work. When a golem is
killed and generates items to drop, it first rolls a random item and
sees whether the golem's material is valid on that item. If so, it will
drop that item instead of its default drop. Other drops are unaffected.

Some special casing is required for balance: glass golems never drop
potions, and paper golems never drop scrolls or spellbooks. (The
appropriate way to do this, if any, would be something like dnethack
spell golems.)

Gold golems also drop rather more zorkmids (they can still generate
other gold items), and the amount of zorkmids is made independent from
Luck.