Fix: wishing for silver dragon scale mail didn't work
On account of "silver" being parsed as a material. The fix is rather
hacky, but I'm not sure of any better way to do it - it'll specifically
check for "silver dragon" before checking the list of materials.

Also contains some cleanup for other things in objnam.c, such as it
still looking for "studded leather armor", etc.