Always show a message when creating monsters
Several medium size refactors are part of this, mostly revolving around
merging various bits of monster-creation logic into calling
create_critters rather than doing their own thing.

create_critters itself is refactored to take the creator monster as an
argument and to return the number of monsters created that the player is
aware of. It is now responsible for printing "[Monsters/A foo] appear[s]
from nowhere!"

One minor loss of detail involved in making the muse create monster
routines call create_critters: monsters won't create amphibious monsters
if they're in water and you're on land (it now depends only on whether
the creator is in water at the time), but it's more symmetric like this
anyway.

Bag of tricks tipping logic is moved from tipcontainer() into
bagotricks(), so that it can create its monsters in one go of
create_critters().

Summon nasties messages are moved from mcastu into nasty(), meaning that
they will also be printed when the Wizard random harassment causes
nasties to be summoned.