Implement leprechaun stealing of all gold items
First, findgold() now takes a second argument, which specifies whether
or not it should only be checking for coins (since a lot of places it is
called assume that it should only find coins, such as shk logic.)
findgold(otmp, TRUE) is equivalent to the previous findgold(otmp).

Then, everywhere that involves uhitm, mhitu, and mhitm leprechaun
gold-stealing, as well as the portions of monster AI that can be
converted to respect gold objects, is updated to use findgold with
FALSE and handle leprechauns stealing gold.