Shop loot and actual gear from inhabitants of Orcish Town
Based on the recent change that allows monsters to be specified as dead
in the des file (dropping their starting inventory and possibly a
corpse), Orcish Town now employs this to produce its loot from
shopkeepers, watchmen, and the priest, instead of poorly duplicating the
makemon code in the des file. Possibly in the future some dead gnomes
and dwarves should be added, but for the moment I'm going to see how it
works out as is.

In addition, the inventory of the other three shops of Frontier Town is
also added in as loot (a 9 tool shop, a 4 deli, and a 6 lighting store.)
Because we can't currently tell the level parser to stock a shop with
items but no shopkeeper, it does an approximation: exact for the tool
shop, decent for the deli (only food items), and poorer but overall more
beneficial to players for the lighting store (a 40/40/10/5/5 split of
tallow candles, wax candles, oil lamps, lanterns, and magic lamps - in
Frontier Town, the magic lamps would only have a 3% chance each to
generate). These are randomly scattered around the town, and of course
are free of charge.

Because shopkeepers commonly carry attack wands and healing potions, I
have placed boulders on top of their corpses - the orcs would otherwise
inevitably get to them first and make the town all that much more
difficult to conquer.