Replace undead on Astral Plane with random A
This is primarily based in flavor: why do the holy priests and angels
tolerate the presence of high-level undead in the high temples of the
gods?

Another reason for this is that the three L that were removed act as
either a spoiler tax or as a punishment for genocideless players: the
odds of the L generating as master or arch-liches were fairly high for a
typical character, so unless L had been genocided, the player would
arrive on Astral and immediately be mobbed by them.

The odds of one or more of the random A being Archons is considered a
suitable replacement for the difficulty of the L and V.