Mjollnir now wakes monsters, can be invoked for lightning bolt
The waking up is because you would expect something that causes
lightning to strike its target to cause a thunderclap. Since this is a
nerf, and also to increase Mjollnir's usefulness as a Valkyrie ranged
weapon, also grant it the invoke power of an 8d6 lightning bolt. This
invoke uses the same timeout rules as any other artifact.