Metallivores usually don't destroy the metal object they eat
Metal ought to be pretty slow-digesting, and the annoyance of a rock
mole cleaning out thousands of zorkmids in a treasure zoo or vault is
well-documented. With this change, metallivores eating off the ground
will with 6/7 chance put the eaten item into their inventory rather than
destroying it. This isn't a perfect solution - it means a player or
other monster could steal the items out of the monster's "stomach", but
as of yet we don't maintain a separate inventory chain for items a
monster has consumed. Similar problem for gelatinous cubes, I guess.