Port SpliceHack / 343-nao hilite_hidden_stairs implementation
This adds the option "hilite_hidden_stairs", which defaults to TRUE and
is settable in-game or via the rc file. When this is enabled, in the
curses interface, the background color of stairs obscured by objects
(but NOT by monsters) will be set to red, and the current foreground
color will be kept.

In the tty interface it only reverses the colors similar to an object
pile, which I consider a bug and will fix.