Fix tty interface hidden stairs hilighting
Makes it consistent with the curses behavior (and also other variants):
hidden stairs are always hilighted with a red background and their
normal foreground. Implementation is mostly copied from dnethack/dNAO.

Architecture-wise, this adds a new function term_start_bgcolor to the
tty library, which is the same as term_start_color but for the
background.