Implement Fourk's random names for desirable weapons
This system generates certain non-artifact weapons with a name randomly
selected from a list of rather macho-sounding strings. Weapons that are
eligible to receive such a name must:
 - Not be an artifact
 - Not be a stack (stackable is ok: currently, projectiles /are/
   allowed, meaning you could step on a dart trap and generate a Dart Of
   Glory; if it's too silly then they can be disqualified later)
 - Be uncursed or blessed
 - Be positively enchanted, with an enchantment higher than a d5 roll.
   If it was randomly made erodeproof, this counts as +1 enchantment in
   this formula.

A weapon that receives a name is thus known to the player to be
desirable because it probably has a good enchantment (and an unspoiled
player might confuse it for an artifact, I suppose).

The actual details of object generation aren't changed; it just means
the player is more likely to notice a random highly enchanted weapon.