Fix: panic in curses mode when looking up string that matched nothing
The logic was a bit incorrect, allowing it to go through the process of
creating a window when a string was provided that didn't match any
monsters, didn't match any objects, and didn't match anything in
data.base. The curses windowport inexplicably panicked when faced with
creating an empty window, but it's still a bug, assuming that the
correct behavior would otherwise be to impossible() or something.

This couldn't happen when the user typed a string. In the report it was
found by the player using the / command to look at himself (with a
character whose name matched nothing).