Fix: weapons in initial inventory could receive weapon onames
Due to mksobj being called with init set to TRUE (needed for certain
types of items, such as cameras), which could cause a weapon to roll a
high enchantment and get a weapon name, but then the weapon would be set
to +0 or +1 or whatever in postprocessing of the created object. Fix
this by removing any onames on specific objects created in starting
inventory.