Remove zruties
This is in preparation for moving sea monsters to z to get them off of a
punctuation mark. z was chosen for a few reasons: it resembles a
squiggly eel-like monster, it's not particularly disruptive because it
was a one-species class, and zruties are so uninteresting of a monster
that I expect few people are actually going to notice they're gone.
Contrary to some people's expectations, referring to the class as "zea
monsters" is not particularly a reason for choosing z.