Move mind flayers and quantum mechanics into U (aberrations)
This resolves a long standing unfairness in the game: the dreaded
mkclass of 'h' resolving to a mind flayer or master mind flayer in the
Mines. They previously didn't have any better class than humanoids to
live in, but opening up a class for aberrations allows them to join it
(noting that they are indeed classified as that in D&D).

Quantum mechanics are being moved off of Q primarily to free it up for
elves (Quendi). But they seem to me to be aberration-y enough to fit
into U without any real consequences of flavor.