Move elves to the Q monster class
The main reason for this is to reduce the overcrowding and color
conflicts of @. Elves are distinct enough from humans to be on a
different glyph (notably, for genocide purposes). Q was recently freed
up for this reason, it being a single-monster class that could work with
some flavor for elves: NetHack elves take after Tolkien, and Tolkienian
elves refer to themselves as Quendi.