Add two Oracle level variants
These are meant to add variety without really changing the balance of
this level. These both feature water rather than a walled subroom. The
"oracle-2" level contains a square ring of water with the fountains
around the corners, and a statue of each type of elemental.  The
"oracle-3" level contains a larger circular ring of water, and only
three fountains inside the ring, but also four trees at the corners of
the ring. It also contains statues of various oracle-type monsters. In
both of the levels, there is a break in the ring so the player can walk
up to the Oracle as normal.

Due to limitations in how REGIONs interact on LEVEL type levels, I
wasn't able to find a way to define a subregion of the large Oracle room
as the "delphi" area. So the slightly worse fallback is to define the
entire large room as that area, meaning the Oracle welcomes you once you
step into it.