Enable grass to be specified in the level compiler
It can be specified in MAPs using the ',' or 'g' characters (the g is
provided in case future level designers decide commas aren't
distinguishable enough). It is added as a full map character, so it can
also be specified in things like INIT_MAP to use a foreground or
background fill of grass.