Redo Monk home level and quest flavor text entirely
It's now a proper monastery, with outer and inner walls, columns,
interior buildings, and a pagoda forest, using the nice new grass
terrain to turn what would otherwise have been an empty field of . tiles
into lawns and pathways.

Quest text relating to this level is also mostly rewritten, primarily
because a lot of the messages were duplicates of Priest quest ones. It's
not yet a full rewrite of this quest's text; that will come with the
changes to the later levels.