Redo Monk locate level entirely
Now is a switchbacky mountain path, ending in a fortress guarded by
xorns and elementals. There are a lot of earth elementals, but unlike
the former level there actually are some other types as well. This level
was formerly just a copy of the Rogue locate level and was bland and
flavorless.

Quest text relating to this level is also rewritten, but that's only two
messages.