Replace Ranger home level entirely
The new level map is basically Khor's Ran-home level, with some 3.6
level parser improvements and a few other tweaks. Notably, the entire
level floor is grass, there is a coaligned altar in the grove, and the
chest contains a guaranteed bow and arrows.

No editing was done for the quest text.