Refactor: reduce has_ceiling to a no-argument function ceiling_exists
This is in preparation for adding a level flag that indicates whether
the level is outdoors. has_ceiling was incompatible with this because
the game only has one level in memory at the time, and it wouldn't be
able to return accurate information if it were called with anything
other than &u.uz since the level flags of that other level would be
inaccessible. Happily, the only calls to has_ceiling were calling it
with &u.uz.