Add "outdoors" level flag, and some initial behavior for it
Behavior at the moment consists of traps that require the ceiling to
make any sense (rust traps and rock traps) not generating. It is also
added to the new ceiling_exists function, which means that effects such
as a scroll of earth doing nothing will now do nothing on
outdoors-flagged levels.

Notably, rust and rock traps can still be specifically asked for in
outdoor levels (say, if one wanted to flavor some area as an overhanging
rock face or the interior of a building) and they will generate - but
they won't generate if a random trap is requested. (Such flavoring might
get weird unless we want to in the future define which squares are and
aren't outside - you'll still throw something "up into the sky" while
inside such a "structure", for instance.)

Every relevant special level has been flagged as outdoors. (All of these
are Quest levels; the rest of them are either Planes which are not
outdoors or are in the dungeon.) Fort Ludios is ambiguous; for now it's
being declared as not outdoors.