Fixes: floodfill panic and wallify levl overflow
I've reported the panic and submitted a pull request for the overflow,
but I don't expect them to get fixed soon enough to power the level
update they rely on. If these get fixed/merged into vanilla later on and
cause merge conflicts, I'll remove the changes in this commit in favor
of the devteam's.