Spice up Barbarian upper filler level with rock outcrops
It's still *primarily* a big open plain, with the same monster and
object distribution, but the stairs are now on opposite ends of the
level as is custom, and there are these big randomly generated hunks of
STONE sitting around. The mesas can generate in such a way that they
block the level vertically, but there's a guaranteed path of . between
the stairs, and the inaccessibles flag is on so that any pockets that
generate have a way out. It's also marked outdoors and hardfloor.

The changes to this level revealed two vanilla crashes that only crop up
on big wall-to-wall open levels like this - if you are cherry-picking
this level, make sure to also get those