Monsters can now hide under grass and some other features
Grass, headstones, altars, sinks, ladders, and thrones all provide
enough cover for a monster that can hide under objects, even if there
are no objects there. They behave exactly the same; for instance, a
concealable monster that steps onto a sink will vanish under it and
become undetected.

The AI for concealable monsters is tweaked so that in large areas of
concealing terrain such as grass fields, they will pick their spot to
move and *then* decide with their 90% chance not to move, *if* the spot
would be moving out of concealment. Previously this was tested before
they actually looked for new places to move, meaning that a monster in a
field of grass would be pointlessly sluggish.