Livelog when the player kills their own pet
This is normally a fairly rare event because of its harsh penalty and
lack of any real benefit, so it seems like it wouldn't be spammy to
livelog it. I may be wrong on this, but there's only one way to find
out.

If you have livelog configured and wish to livelog this target, make
sure your LIVELOGFLAGS in the sysconf are not set to the default 0xFF,
because the flag for this livelog is 0x0100.