Prevent monsters generated in rooms from spawning on staircases
The main goal of this is to make it impossible for the player to enter a
randomly generated level where a swarm of monsters happened to generate
on the stairs and is pushed one or two spaces away with the stairs being
occupied. This will still happen in bones levels if something was on the
stairs when the previous player died, though.