Correct formatting of death when killed stuck in creature form
Previously any such death would list the reason as "killed while stuck
in creature form", which loses a lot of detail. In particular, any game
which was sentient_arised that ended in death would have registered as
this.

To circumvent this, rehumanize() in its done(DIED) codepath no longer
sets that as the killer name. This involved adding a lot of manual
setting of killer.name wherever rehumanize might be called, unless it
had checked unchanging prior to calling it. Largely this was just
copying the regular killer, but there are two exceptions: polyform
instadeaths (iron golem hit with a rust attack, e exploding itself)
which needed new death messages, and the "monster killed player" code,
which needed refactoring to extract the killer formatting component of
done_in_by.

Initially I wanted to add this to the code that formats the death
reason, but that code that adds ", while [helpless]" to the death seems
to be dead; I'll try to check with the devteam on that. For now, the
actual changes are implemented in the while= field of the xlogfile, and
rather than generic "creature form" it uses whatever your polyform
actually is.
 * while=sleeping
 * while=stuck in hobbit form
 * while=sleeping and stuck in hobbit form