Add polyinit mode
This is a non-scoring mode like explore mode that causes you to begin
the game permanently polymorphed into a monster. It already exists in
NetHack4 and its variants; to my knowledge, this is the first
implementation in a 3.x variant.

A side effect of this feature is a behavioral change; if you are killed
in a polyform with any unchanging at all, you will die. The code that
previously didn't kill you if your HP was positive is removed.
Practically, the only thing this will affect is that if you died in some
way like burning to a crisp as a paper golem or being exploded as a gas
spore while _not_ in polyinit mode but wearing unchanging, you will now
die where previously you would rehumanize.