Polyinit: allow illegal polyforms, but not mplayers and uniques
Due to popular demand for Archon polyinit. Now polyinitting to something
which is non-polymorphable works, but a large warning message is printed
at the start of the game saying that you should expect this to break
stuff. Officially, only polyinits to legal polyforms are supported.

Polyinitting to uniques and player monsters would probably REALLY break
stuff, so that's still forbidden.

There's currently one known crash with Archon polyinit; a patch is
available but I haven't applied it yet.