Refactor: generalize silver hating to all materials
With the intention of adding monster behavior that hates things that are
not silver. This attempts to be as material-agnostic as possible with
regard to the other material. Hate_silver is now Hate_material(),
mon_hates_silver is now mon_hates_material, and so on.

However, silver will still enjoy the most dramatic effects of any
material: it still sears flesh and does d20 damage, whereas all the new
default "some other material vs. a monster" code results in d6 damage
and messages about "[monster] flinches" or "it hurts to touch". There
aren't actually that many places where a distinction has to be drawn,
however.