Rebalance object material weights
There were various points of dissatisfaction I had with the
trying-to-be-as-realistic-as-possible system. This abandons trying to
stick to reality and uses arbitrary units of density, which will
hopefully make for balanced gameplay while still feeling realistic.

Flavor changes for e.g. why a gold sword no longer weighs 2.5 times the
weight of an iron sword can be handwaved as gilding/silvering/etc,
different thickness/volume of the shape, and so on.