Replace the Barbarian goal level
No longer a duplicate of the awful Mimic of the Mines level. Now
consists of a small abandoned village, with the center barricaded off by
boulders, where Thoth Amon has made his hideaway.

His horde is larger now than it was. Once the player enters the level,
all the marauding enemies outside the village will charge them. There's
a line of burning candles (we lack campfires) on the ground by which the
player can see their approach. Within the village are more elite
monsters.

There is no longer an altar, but the sheer amount of loot from the horde
should probably be enough of a tradeoff.