Port F growing on old corpses from SLASH'EM
Behavior is more or less the same; messages are tweaked a bit. Some
architectural details differ between SLASH'EM and 3.6, but that
shouldn't really be noticeable.

Using SLASH'EM's formula, a regular newly killed corpse will have about
a 50% chance to turn into a random F, on a random turn between turning
maybe-tainted and rotting away completely. If the random F chosen can
move, it tries once more to pick a random F and uses that whether or not
it can move, giving a bias towards stationary fungi.

Corpses from the traps on early dungeon levels, as far as I know, are
fresh and I assume some will start decaying into F once the hero loads
that level. If the F selected can move, I suppose this opens up the tiny
possibility of a trap on early levels not being instantly recognizable
by its corpse, but this would already be the case after the corpses
decayed anyway.