Melee combat wakes up nearby monsters
Another feature ported from dnethack, with some more changes (the
formula for the amount of noise produced is now used as a direct radius
rather than square-rooting it; there is a 1/10 chance that an otherwise
noisy attack produces no noise at all, upped to 1/5 for a stealthy
player). Non-melee combat does not wake anything up.

The purpose of this change is to address the flavor problem of a
stealthy player methodically slaughtering one of the many types of
special rooms that is packed with sleeping monsters, one by one. It will
therefore make clearing out those rooms more bothersome if melee is
used, and hopefully will diversify player strategies a bit.

Introduces a couple macros from dnethack to simplify the code involved.