Multiline, dramatic appearance messages for bosses
Ultimately this is from SpliceHack, though I managed to get it working
by using STRAT_APPEARMSG instead of adding another bit to the monst
struct.

The text entries are placed in quest.txt as non-role-specific chunks,
and are com_pagered into being the first time the hero encounters the
corresponding monster. The hero generally needs to be able to spot the
monster, and in some cases also needs to be able to see the literal
space where it's standing, though there are some exceptions.

Bosses that get appearance messages include all demon lords, the Riders,
the Wizard, and Vlad. Medusa was discussed but ultimately thought to be
irrelevant, just because the player's interaction with her usually only
takes one turn.

I tried my best to ensure that a boss would always get a message on
their first appearance at an appropriate time, but there are probably
still some bugs in the exact logic of it.