Port FIQHack's ring initial enchantment rules
The main math uses rne(2), which as per the previous commit can now
occasionally generate very high values. The sign of the enchantment is
usually positive if blessed, usually negative if cursed, and positive
about twice as often as negative if uncursed. This is the one change
from FIQHack's formula: in an effort to avoid lots of very negative
useless uncursed rings, more of them will be positive than before (which
in FIQHack is about 55/45.)

Any ring can end up +0, but only 2% of the time.